feat(git): apply pgp sig event tag to commit
to maintain correct commit ids which is required to apply multiple
commits

its noted that no tests are written and the scenario where the author
and committer differ has not been tested

clearly the validate_patch_applied function has code that corrects an
error where the author / committer 'signatures' do not apply correctly.
this will not be fixed under a pgp signed commit scenario.