fix: intermittent crashing by removing block_on
a crashing bug was identifed when users ran ngit with large relay sets.

the test suite would also stop at random tests and produce a
`...running for over 60 seconds` error but only on nix configuration
and not when using rustup. this change fixes this so the ngit must have
been crashing more often when ran with the reduced resources of a nix
shell.