feat(remote): fallback to ssh or https
ssh clone urls are problematic as it requires authentication for
fetching if users don't ave ssh keys setup for authentication with
the git server. so cloning / fetching will fail.

some users only authenticate using https and some only using ssh
so we cant just use https for everything.

this commit falls back to the other protocol if the first one failed
for list, fetch or push