fix(remote): support lightweight tags
I have now replicated the issue discussed in the last commit by
overwriting my global git config item tag.gpgSign and setting
it back to false, which is default the default.

ngit was only supporting annotated tags and fiatjaf was pushing
a lightweight tag.

I'm confident that this will resolve the issue