feat(send): support PR and PR update events
send as a PR if the commit would make patches that are too big for
nostr events.

send as a PR update if the proposal is PR.

send as a PR, revising a patch root, if patches would be too big.

in tests `get_pretend_proposal_root_event` has to be a actual proposal
with a tip, rather than just a cover letter, so we have replaced it.